requirements.txt
PyGithub==1.26.0
rethinkdb==2.3.0.post3
